chore(AllClass): 패키지 분리
