+++ src/main/java/blackjack/domain/card/Deck.java
@@ -1,6 +1,5 @@
-package blackjack.domain;
+package blackjack.domain.card;
 
-import java.util.Collections;
 import java.util.List;
 import java.util.Stack;
