+++ src/main/java/blackjack/domain/card/Number.java
@@ -1,8 +1,7 @@
-package blackjack.domain;
+package blackjack.domain.card;
 
 import java.util.Arrays;
 import java.util.List;
-import java.util.stream.Collectors;
 
 public enum Number {
     ACE(11),
