+++ src/main/java/blackjack/domain/card/Symbol.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.card;
 
 import java.util.Arrays;
 import java.util.List;
