+++ src/main/java/blackjack/domain/game/Result.java
@@ -1,4 +1,7 @@
-package blackjack.domain;
+package blackjack.domain.game;
+
+import blackjack.domain.participant.Dealer;
+import blackjack.domain.participant.Player;
 
 public enum Result {
     WIN,
