+++ src/test/java/blackjack/domain/participant/PlayersTest.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.participant;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
